requirement.txt
pandas-ta
beautifulsoup4
plotly
